Oraganize tests.
