+++ src/main/java/com/github/dakusui/actionunit/core/context/ContextFunction.java
@@ -34,8 +34,8 @@ public interface ContextFunction<R> extends Function<Context, R>, Printable {
         ));
   }
 
-  static <R> ContextFunction<R> of(Supplier<String> formattter, Function<Context, R> function) {
-    return new ContextFunction.Impl<R>(formattter, function);
+  static <R> ContextFunction<R> of(Supplier<String> formatter, Function<Context, R> function) {
+    return new ContextFunction.Impl<R>(formatter, function);
   }
 
   class Impl<R> extends PrintableFunction<Context, R> implements ContextFunction<R> {
